UD done with id

+++ car/views.py
@@ -6,7 +6,8 @@ from rest_framework.response import Response
 from rest_framework import viewsets
 from rest_framework import status
 from .models import Car, Cancel, Feedback, Booked
-from .serializers import ShowCarSerializer, CarRegisterSerializer, BookCarSerializer, FeedbackSerializer, CancelSerializer
+from .serializers import ShowCarSerializer, CarRegisterSerializer, BookCarSerializer, FeedbackSerializer, \
+    CancelSerializer
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from django_filters import rest_framework as filters
 from django_filters.rest_framework import DjangoFilterBackend
@@ -23,7 +24,7 @@ class CarFilter(filters.FilterSet):
 
     class Meta:
         model = Car
-        fields = {                      # iexact
+        fields = {  # iexact
             'company': ['icontains'],
             'car_model_name': ['icontains'],
             'city': ['icontains'],
@@ -53,17 +54,24 @@ class CarRegister(APIView):
         if serializer.is_valid():
             car = serializer.save(ownerId=user)
             if car:
-                return Response({'id': car.id, 'number_plate': car.number_plate, 'model': car.car_model_name, 'message' : 'Car Registered'}, status=status.HTTP_201_CREATED)
+                return Response({'id': car.id, 'number_plate': car.number_plate, 'model': car.car_model_name,
+                                 'message': 'Car Registered'}, status=status.HTTP_201_CREATED)
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def put(self, request):
-        car = request.car
+
+class CarUpdate(APIView):
+    serializer_class = CarRegisterSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def put(self, request, car):  # NOTE: car is a object not id
+        # car = request.car
+        user = request.user
         try:
-            u = Car.objects.get(id=car.id)
+            u = Car.objects.get(id=car)
         except Car.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
-        if request.method == "PUT":
+        if request.method == "PUT" and user == u.ownerId:
             serializer = CarRegisterSerializer(u, data=request.data)
             data = {}
             if serializer.is_valid():
@@ -71,15 +79,18 @@ class CarRegister(APIView):
                 data['success'] = "update successful"
                 return Response(data=data)
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({'error': 'access denied'}, status=status.HTTP_401_UNAUTHORIZED)
 
-    def delete(self, request):
-        car = request.car
+    def delete(self, request, car):  # NOTE: car is a object not id
+        # car = request.car
+        user = request.user
         try:
-            u = Car.objects.get(id=car.id)
+            u = Car.objects.get(id=car)
         except Car.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
 
-        if request.method == "DELETE":
+        if request.method == "DELETE" and user == u.ownerId:
             operation = u.delete()
             data = {}
             if operation:
@@ -87,21 +98,49 @@ class CarRegister(APIView):
             else:
                 data["failure"] = "delete failed"
             return Response(data=data)
+        else:
+            return Response({'error': 'access denied'}, status=status.HTTP_401_UNAUTHORIZED)
 
 
-class BookCar(APIView):
+# class BookCar(APIView):  # Do it with this way or just pass car ID and delete that car
+#     serializer_class = BookCarSerializer
+#     permission_classes = (IsAuthenticated,)
+#
+#     def post(self, request):
+#         user = request.user
+#         serializer = self.serializer_class(data=request.data)
+#         if serializer.is_valid():
+#             car = serializer.save(renterId=user, booked_on=timezone.now().date())
+#             if car:
+#                 return Response({'car id': car.id, 'message': 'Car Booked '}, status=status.HTTP_201_CREATED)
+#
+#         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class BookCar(APIView):  # Do it with this way or just pass car ID and delete that car
     serializer_class = BookCarSerializer
     permission_classes = (IsAuthenticated,)
 
-    def post(self, request):
+    def post(self, request, carid):
         user = request.user
+        try:
+            u = Car.objects.get(id=carid)
+        except Car.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         serializer = self.serializer_class(data=request.data)
-        if serializer.is_valid():
-            car = serializer.save(renterId=user, booked_on=timezone.now().date())
-            if car:
-                return Response({'car id': car.id, 'message': 'Car Booked '}, status=status.HTTP_201_CREATED)
+        try:
+            Booked.objects.get(carId=carid)
+            return Response({'message': 'Car is already Booked by someone else. Sorry!'},
+                            status=status.HTTP_400_BAD_REQUEST)
+        except:
+            if serializer.is_valid():
+                car = serializer.save(renterId=user, booked_on=timezone.now().date(), carId=u, company=u.company,
+                                      car_model_name=u.car_model_name, ownerId=u.ownerId, address=u.pickup_address,
+                                      owner_phone=u.ownerId.phone, renter_phone=user.phone)
+                if car:
+                    return Response({'carid': car.id, 'message': 'Car Booked '}, status=status.HTTP_201_CREATED)
 
-        return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class FeedbackCar(APIView):
@@ -122,16 +161,16 @@ class CancelBookedCar(APIView):
     serializer_class = CancelSerializer
     permission_classes = (IsAuthenticated,)
 
-    def post(self, request):
+    def delete(self, request, bookid):  # Do it with this way or just pass booking ID and delete that car
         user = request.user
-        Id = request.POST['carId']
         serializer = self.serializer_class(data=request.data)
         # bookedcar = Booked.objects.filter(carId=Id)
-        if serializer.is_valid() and Id:
-            Booked.objects.filter(carId=Id).delete()
-            car = serializer.save(renterId=user, cancelled_on= timezone.now().date())
+        if serializer.is_valid():
+            b = Booked.objects.filter(id=bookid)
+            Booked.objects.filter(id=bookid).delete()
+            car = serializer.save(renterId=user, cancelled_on=timezone.now().date())
             if car:
-                return Response({'Booking canceled.'}, status=status.HTTP_201_CREATED)
+                return Response({'message': 'Booking canceled.'}, status=status.HTTP_204_NO_CONTENT)
 
         return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
@@ -174,10 +213,9 @@ class MyRentedCarList(ListAPIView):
 #     ordering = ('seats', )
 #     search_fields = ('car_type', 'rent', 'number_plate')
 
-    # filterset_class = CarFilter
-    #
-    # def carfilter(self, request):
-    #     car = request.get_queryset().order_by('company').last()
-    #     serializer = request.get_serializer_class()(HomeAndFilter)
-    #     return Response(serializer.data)
-
+# filterset_class = CarFilter
+#
+# def carfilter(self, request):
+#     car = request.get_queryset().order_by('company').last()
+#     serializer = request.get_serializer_class()(HomeAndFilter)
+#     return Response(serializer.data)
